Rewrite command channel.
Bench: 4496389

+++ src/engine.rs
@@ -173,7 +173,7 @@ impl Engine {
                     return;
                 }
 
-                let Ok(val) = value.parse::<usize>() else {
+                let Ok(val) = value.parse::<u32>() else {
                     println!("info string Unknown value {value}");
                     return;
                 };
